Finally got the tag drawing exactly how I want it
We now only round the corners on the start of a tag and on the end of a tag if it starts and ends on that line. We use a rounded rectangle path instead of a rounded rectangle to do so, Canvas.drawRoundRect() doesn't support defining specific corners' radii, and Path.addRoundRect() does. We also don't bother trying to draw on lines which are past our max.