Stop cards from flashing when returning from info activity
The changes made here actually mean that changing the rating while in the info activity won't cause it to be updated when returning to the list.

If we also update the tags though, it will be, since for some reason those updates always go through.

Hopefully a nice way of fixing this will be presented in the future, but I'm not so worried about it for now.