It is actually mostly working
You can rotate all you want, and it will still work as long as the orientation type is the same. But if you go from P-->L or L-->P and then go back, it doesn't work. I assume this is because it keeps some state around which is based on the bounds which were used? Idk.