Removed Guava dependency
To do this, we extracted the hashing functionality which we used from Guava's hashing classes. I've made sure to retain the file header license statements, as well as disclaim any credit for the code and provide a link to the license file in the Guava repo in the javadoc comments on the classes.